Use abbreviated month names in all signup forms to save space

+++ src/pages/SIGNUP- NFP.owt61.js
@@ -43,7 +43,7 @@ async function populateDateRepeater() {
 					month: dateObj.getMonth(),
 					year: dateObj.getFullYear(),
 					day: dateObj.getDate(),
-					monthName: dateObj.toLocaleDateString('en-US', { month: 'long' })
+					monthName: dateObj.toLocaleDateString('en-US', { month: 'short' })
 				};
 			})
 			.sort((a, b) => a.date - b.date);
